modifiche richieste

+++ src/main/java/it/polito/ai/laboratorio3/dtos/EssayDTO.java
@@ -10,7 +10,7 @@ public class EssayDTO {
     public enum stati {Letto, Consegnato, Rivisto, Terminato}
 
     private Long id;
-    private int voto;
+    private Long voto;
     public stati stato;
     private String idStudente;
